drt 4324 switch back to sbt .11, introduce SplitRatios case class
