+++ server/src/main/scala/services/workloadcalculator/WorkloadCalculator.scala
@@ -3,7 +3,7 @@ package services.workloadcalculator
 import org.joda.time.{DateTime, DateTimeZone}
 import org.slf4j.LoggerFactory
 import spatutorial.shared.FlightsApi.{QueueName, QueuePaxAndWorkLoads}
-import spatutorial.shared.SplitRatios.SplitRatios
+import spatutorial.shared.SplitRatiosNs.SplitRatios
 import spatutorial.shared._
 
 import scala.List
@@ -87,7 +87,7 @@ object PaxLoadCalculator {
 
   def voyagePaxSplitsFlowOverTime(splitsRatioProvider: (ApiFlight) => Option[SplitRatios])(flight: ApiFlight): IndexedSeq[(MillisSinceEpoch, PaxTypeAndQueueCount)] = {
     val timesMin = new DateTime(flight.SchDT, DateTimeZone.UTC).getMillis
-    val splits = splitsRatioProvider(flight).get
+    val splits = splitsRatioProvider(flight).get.splits
     val splitsOverTime: IndexedSeq[(MillisSinceEpoch, PaxTypeAndQueueCount)] = minutesForHours(timesMin, 1)
       .zip(paxDeparturesPerMinutes(if (flight.ActPax > 0) flight.ActPax else flight.MaxPax, paxOffFlowRate))
       .flatMap {
