+++ shared/src/main/scala/spatutorial/shared/SplitRatiosNs.scala
@@ -1,0 +1,8 @@
+package spatutorial.shared
+
+object SplitRatiosNs {
+//  type SplitRatios = List[SplitRatio]
+  case class SplitRatios(splits: List[SplitRatio])
+
+  case class SplitRatio(paxType: PaxTypeAndQueue, ratio: Double)
+}
